Scheduling containers among a pool of servers.

+++ omgservers-service/src/main/java/com/omgservers/service/operation/getConfig/ServiceConfig.java
@@ -116,9 +116,9 @@ public interface ServiceConfig {
 
         String dockerNetwork();
 
-        int defaultCpuLimit();
+        long defaultCpuLimit();
 
-        int defaultMemoryLimit();
+        long defaultMemoryLimit();
     }
 
     interface InitializationConfig {
